[깃허브 연동, 패키지 리팩터링]
1. 깃허브 연동을 프론트와 협업을 위한 형식으로  수정
2. 패키지 이름과 위치를 리팩터링
