--- src/main/java/gitME/auth/dto/KakaoDTO.java
@@ -1,13 +1,0 @@
-package gitME.auth.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class KakaoDTO {
-
-    private long id;
-    private String nickname;
-
-}
