+++ src/main/java/gitME/database/repository/CodeStackRepository.java
@@ -1,0 +1,7 @@
+package gitME.database.repository;
+
+import gitME.database.entity.CodeStack;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CodeStackRepository extends JpaRepository<CodeStack, Integer> {
+}
