+++ src/main/java/gitME/database/repository/ExternalLinkRepository.java
@@ -1,0 +1,7 @@
+package gitME.database.repository;
+
+import gitME.database.entity.ExternalLink;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ExternalLinkRepository extends JpaRepository<ExternalLink, Integer> {
+}
