+++ src/main/java/gitME/database/repository/GithubUserRepository.java
@@ -1,0 +1,7 @@
+package gitME.database.repository;
+
+import gitME.database.entity.GithubUser;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GithubUserRepository extends JpaRepository<GithubUser, Integer> {
+}
