+++ src/main/java/gitME/database/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package gitME.database.repository;
+
+import gitME.database.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
+}
