--- src/main/java/gitME/entity/ExternalLink.java
@@ -1,16 +1,0 @@
-package gitME.entity;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-@Entity
-@Data
-public class ExternalLink {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idx;
-    private int userIdx;
-    private String url;
-    private String description;
-}
