+++ src/main/java/gitME/kakao/kakaoAuth/dto/KakaoDTO.java
@@ -1,0 +1,13 @@
+package gitME.kakao.kakaoAuth.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class  KakaoDTO {
+
+    private long id;
+    private String nickname;
+
+}
