+++ src/main/java/gitME/kakao/ss.java
@@ -1,0 +1,8 @@
+package gitME.kakao;
+
+/**
+ * Class ss.
+ * Created by 82105 on 2023-11-14.
+ */
+public class ss {
+}
