--- src/main/java/gitME/repository/CodeStackRepository.java
@@ -1,7 +1,0 @@
-package gitME.repository;
-
-import gitME.entity.CodeStack;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CodeStackRepository extends JpaRepository<CodeStack, Integer> {
-}
