--- src/main/java/gitME/repository/ExternalLinkRepository.java
@@ -1,7 +1,0 @@
-package gitME.repository;
-
-import gitME.entity.ExternalLink;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ExternalLinkRepository extends JpaRepository<ExternalLink, Integer> {
-}
