--- src/main/java/gitME/repository/GithubUserRepository.java
@@ -1,7 +1,0 @@
-package gitME.repository;
-
-import gitME.entity.GithubUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GithubUserRepository extends JpaRepository<GithubUser, Integer> {
-}
