--- src/main/java/gitME/repository/RepositoryRepository.java
@@ -1,7 +1,0 @@
-package gitME.repository;
-
-import gitME.entity.Repository;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RepositoryRepository extends JpaRepository<Repository, Integer> {
-}
