--- src/main/java/gitME/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package gitME.repository;
-
-import gitME.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-}
