--- src/main/resources/templates/gitInterlink.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <title>gitInterlink</title>
-    </head>
-    <body>
-        <h1>Hello, gitInterlink test success </h1>
-        <ul>
-            <li th:each="repo : ${repositories}">
-                <span th:text="${repo.name}"></span>
-                <span th:text="${repo.owner.login}"></span>
-            </li>
-            <li th:each="lang : ${langList}">
-                <span th:text="${lang}"></span>
-            </li>
-        </ul>
-    </body>
-</html>
