--- src/main/resources/templates/home.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <title>gitME home</title>
-    </head>
-    <body>
-        <h1>gitME home</h1>
-
-        <!-- 권한 많은 -->
-<!--        <a th:href="@{'https://github.com/login/oauth/authorize?client_id=' + ${githubClientId} + '&scope=repo'}">깃허브 연동 private repo</a>-->
-
-        <!-- 권한 부족한 -->
-         <a th:href="@{'https://github.com/login/oauth/authorize?client_id=' + ${githubClientId}}">깃허브 연동 public repo</a>
-
-
-    </body>
-</html>
